Add Glacier support via Admin Controls for testing
A REST route was added for manipulating S3 resources into various
lifecycle states trypically controlled via Amazon-side policies. This
REST API can be used when testing behavior of your application when an S3
resource has been moved to glacier, expired, in the process of being
restored, etc.

In addition, files remaining in the filestore are loaded upon starting of
the server.