[AR-4152] Add lvh.me as one of the accepted hostnames
When parsing the bucket name from a request, it is optionally in the path
or in the subdomain. For whatever reason, the fake-s3 gem needs to know a
predefined set of hostnames from which to receive a request so it can
correctly determin if there is a subdomain present. 'lvh.me' is
equivalent to 'localhost' for many developer machines, thus it is being
added to this list.