Support delete of bucket
When deleting a bucket, an error would occur due to an XML parser failing
to read an empty body. Turns out, the delete bucket command was being
treated as a 'delete objects' command. That command expects there to be
an XML body containing the objects to delete.

The logic for determining between a bucket delete and a set of objects
delete was improved. In addition, when deleting a bucket that does not
exist, do not raise an error of "bucket does not exist".